Replace tagName in CSS
