+++ scripts/build.js
@@ -1,41 +1,26 @@
+/**
+ * Reads files in `node_modules/@vaadin/**`
+ * Writes them to `packages/vaadin/**`
+ * Performs replacement tagNames: `vaadin-` => `vaadin{majorVersion}-`
+ */
+
 import { glob } from "glob";
 import Path from "path";
 import fs from "fs";
 import { init, parse } from "es-module-lexer";
 import { transformImports } from "./transformModuleImportsPlugin.js";
-import { elementMeta } from "./element-meta.js";
-import { versionMeta } from "../version.js";
 import { ignorePackages } from "./ignore-packages.js";
-import { supplementalElementNames } from "./supplemental-element-names.js";
+import {
+  processTagNames,
+  processTagNamesNaive,
+  processLocalName,
+} from "./replacement-helpers.js";
+import { allElementNames, allEventNames, allPackageNames } from "./meta.js";
+import { versionMeta } from "../version.js";
 
+export const majorVersion = versionMeta.vaadinVersion;
 const nodePackagesRoot = "node_modules/@vaadin";
 const localPackagesRoot = "packages/vaadin";
-export const majorVersion = versionMeta.vaadinVersion;
-
-function escapeRegExp(text) {
-  return text.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&");
-}
-
-export function allElementNames() {
-  let accumulator = [];
-  elementMeta.forEach((meta) => {
-    accumulator = [...accumulator, ...meta.elementNames];
-  });
-  accumulator = [...accumulator, ...supplementalElementNames];
-  return [...new Set(accumulator)].sort();
-}
-
-function allEventNames() {
-  let accumulator = [];
-  elementMeta.forEach((meta) => {
-    accumulator = [...accumulator, ...meta.eventNames];
-  });
-  return [...new Set(accumulator)].sort();
-}
-
-function allPackageNames() {
-  return elementMeta.map((meta) => meta.package);
-}
 
 function findPackages(dir) {
   const ignore = new Set(ignorePackages);
@@ -175,106 +160,6 @@ function processCustomElementsRegistry(content, filePath) {
   return cleaned;
 }
 
-function computeLiteralRe() {
-  // matches string literals like "foo-bar", 'foo-bar', `foo-bar`
-  // added commas to the mix because a couple of places in the code will try to search for
-  // tag names like so:
-  // this._setInputElement(this.querySelector("vaadin-text-field,.input"));
-  // could technically add optional commas within the quote maches, but will instead just
-  // include them where the quotes are matched
-  const names = allElementNames()
-    .map((name) => escapeRegExp(name))
-    .join("|");
-  return new RegExp(`[\`'",](${names})[\`'",]`, "g");
-}
-
-function computeTagRe() {
-  // tries to match opening and closing tags in markup
-  // eg:  <foo-bar>, </foo-bar>
-  const names = allElementNames().join("|");
-  return new RegExp(`([<]|<\\/)(${names})`, "g");
-}
-
-function computeUndoEventsRe() {
-  // tries to rename event-name literals that were over-aggressively renamed
-  // due to they align with a tagName
-  // note: this may nonger be needed
-  const allTags = new Set(allElementNames());
-  const names = allEventNames()
-    .filter((value) => value.indexOf("vaadin-") > -1)
-    .filter((value) => !allTags.has(value)) // if it _is_ a tag name, use special handling
-    .map((value) => value.replace(`vaadin-`, `vaadin${majorVersion}-`))
-    .join("|");
-  return new RegExp(`${names}`, "g");
-}
-
-// tries to rename events that were renamed because they align with a tagName
-// note: this may nonger be needed
-function computeUndoEventsStrictRe() {
-  const allTags = new Set(allElementNames());
-  const names = allEventNames()
-    .filter((value) => value.indexOf("vaadin-") > -1)
-    .filter((value) => allTags.has(value))
-    .map((value) => value.replace(`vaadin-`, `vaadin${majorVersion}-`))
-    .join("|");
-  return new RegExp(`Event\\(['"\`](${names})`, "g");
-}
-
-const literalRe = computeLiteralRe();
-const tagRe = computeTagRe();
-const undoEventsRe = computeUndoEventsRe();
-const undoEventsStrictRe = computeUndoEventsStrictRe();
-
-function processLocalName(content) {
-  // Usages of `localName` in code are tricky to contend with...
-
-  // example: vaadin-combo-box-overlay creates some custom properties via localName... would perhaps be better
-  // to just change the consumption of those properties (css) so that the version-adorned
-  // names are used, but crrently feel it's better to place along nicely w/ the
-  // standard-fare custom property names
-  // we want this:  --_vaadin-time-picker-overlay-default-width
-  // not this:      --_vaadin24-time-picker-overlay-default-width
-  return content.replaceAll(
-    `const propPrefix = this.localName;`,
-    `const propPrefix = this.localName.replace('vaadin${majorVersion}', 'vaadin');`
-  );
-}
-
-/**
- * Attempts to find HTMLElement tag-names and replace them, eg:
- *
- *   "vaadin-button" -> "vaadin24-button"
- *
- * This is done using string search-and-replace only. There is no attempt to
- * parse the input as codeF
- *
- * @param {string} content : ;
- * @param {Path} filePath
- */
-export function processTagNames(content, filePath) {
-  let result = content;
-
-  result = processLocalName(result);
-
-  result = result.replace(literalRe, (matched) => {
-    return matched.replaceAll(`vaadin-`, `vaadin${majorVersion}-`);
-  });
-
-  result = result.replace(tagRe, (matched) => {
-    return matched.replaceAll(`vaadin-`, `vaadin${majorVersion}-`);
-  });
-
-  result = result.replace(undoEventsRe, (matched) => {
-    return matched.replaceAll(`vaadin${majorVersion}-`, `vaadin-`);
-  });
-
-  result = result.replace(undoEventsStrictRe, (matched) => {
-    return matched.replaceAll(`vaadin${majorVersion}-`, `vaadin-`);
-  });
-
-  return result;
-}
-
 /**
  *
  * @param {string} content : ;
@@ -282,7 +167,7 @@ export function processTagNames(content, filePath) {
  * @returns
  */
 export async function processJs(content, filePath) {
-  const cleanedTagNames = processTagNames(content, filePath);
+  const cleanedTagNames = processTagNames(content);
 
   await init;
   const [imports, _exports] = await parse(cleanedTagNames, filePath.base);
@@ -316,7 +201,7 @@ export async function processJs(content, filePath) {
  */
 function processReadmeMd(content, filePath) {
   let result = content;
-  result = processTagNames(result, filePath);
+  result = processTagNames(result);
   result = replaceNpmScope(result);
 
   const vaadinPackageName = filePath.dir.replace("node_modules/", "");
